chore(Layout): apply new layout structure to pages
A new layout structure where the page title is no longer in the link
header but instead is a H1 tag and the titles of the actual guides have
been changed to make it more readable.

The styles have been updated to hide the H1 tag from the nav menu as
this isn't considered a page section but rather the page title so
doesn't need linking back to.