fix(HowToGuides): update adding the UnityXRCameraRig to 2018.3 version
As VRTK now has a minimum requirement of Unity 2018.3 then the guides
need to be tested in at least this version and updated accordingly.

The image assets have been redone as the UI of Unity 2018.3 has been
redesigned so it looks different.

Any reference to Unity 3d has been changed to Unity and where Unity
is used as an adjective not a noun.