feat(HowToGuides): how to use Unity Button Action
A How To Guide of how to use the Unity Button Action in a simple
form to tie up a basic action of toggling the active state of another
scene GameObject.

It covers how to use the built in Unity controller button mapping
prefabs included with VRTK as well as creating a custom Unity Button
Action.