chore(styles): add underline to section heading 2 items
To better split up the content, border bottom lines have been added
to the h2 tags in the section content to provide better understanding
of where the new sections start.

Also, the oriented text has been removed from the headers on the
main README as it didn't make much sense and messed up the menu items.