fix(Tutorials): account for offsets when reporting velocities
The velocity and angular velocity both require offsetting against
existing GameObject rotations to ensure they report the correct
velocity.

This fix showcases the fix as part of the tutorial to provide an
extra lesson, whilst the commits on the actual tutorial repo are
later on as they are fixes to a previous issue.